+++ react/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @superflag/super-ipc-react
 
+## 0.0.20
+
+### Patch Changes
+
+- update readmes
+- add global electronApi type for advanced usages
+- improve error handling
+- add app version support to the electron bridge and react hooks
+- Updated dependencies
+  - @superflag/super-ipc-core@0.0.20
+
 ## 0.0.19
 
 ### Patch Changes